requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
notion-client==2.4.0
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.14.1
